Improve stream handling. Fix create_mpu_object() to not accept open streams.
Streams (open resources) do not need to support seek or size if the
length option or Content-Length header is set. Streaming alway begins
from the current position for the set length of bytes, unless seekTo is
given.
If length is not given, the stream has to support fstat() for size and
position (when not using seekTo), and is read until EOF.