Bug fix for the NULL byte bug fix in version 1.5.7
Some time ago, I reported a bug with the NULL byte handling in DynamoDB, which specially occurs when saving session data (see https://forums.aws.amazon.com/thread.jspa?threadID=94935).

Now I had the time to check the bug fix and it doesn't work. Contents with NULL bytes are now JSON encoded and can be saved now (with the prefix "json_encoded:"), but the response isn't decoded... So the session data still cannot be deserialized correctly and the sessions are invalid.

My change fixes this problem for the session handler, but the decoding of the response should be done globally, because the encoding for contents with NULL bytes is also done globally.

BTW: The test for this bug also doesn't cover the decoding of the content.